Move functions that check the trade balance to check_balance module * Move check_my_coin_balance_for_swap, check_other_coin_balance_for_swap, check_base_coin_balance_for_swap
